+++ manageNote/src/main/java/com/example/manageNote/controller/NoteController.java
@@ -12,13 +12,20 @@ import com.fasterxml.jackson.databind.cfg.CoercionInputShape;
 import com.fasterxml.jackson.databind.type.LogicalType;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.json.JSONException;
+import org.json.JSONObject;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.ConstraintViolation;
+import javax.validation.Validation;
+import javax.validation.Validator;
+import javax.validation.ValidatorFactory;
 import java.util.Date;
 import java.util.List;
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicInteger;
 
 
 /**
@@ -37,13 +44,14 @@ public class NoteController {
     private static String portPatient = ":8081";
     private static String endpointPatient = "/patient";
 
+    private final Validator validator;
 
-    //L'injection de dépendence peut se faire de pluisuers façons :
-    //La première est en utilisant @Autowired
-    //La deuxième est en mettant en paramètre dans le constructeur
+
     public NoteController(INoteService noteService, ConfDockerService confDockerService) {
         this.noteService = noteService;
         this.confDockerService = confDockerService;
+        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+        validator = factory.getValidator();
 
         mapper = new ObjectMapper();
         mapper.registerModule(new JavaTimeModule());
@@ -87,14 +95,12 @@ public class NoteController {
     public ResponseEntity<Note> find(@PathVariable Integer id){
         Note noteFromDB = noteService.findNoteById(id).orElse(null);
 
-        //S'il n'éxiste pas, envoie statut 404
         if(noteFromDB == null) {
-            //log.error("Error : id Note doesn't exist in the Data Base.");
+            log.error("Error : id Note doesn't exist in the Data Base.");
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
-        //log.info("Returning the note's informations");
-        //Sinon, retourner note
+        log.info("Returning the note's informations");
         return new ResponseEntity<>(noteFromDB, HttpStatus.OK);
     }
 
@@ -108,26 +114,23 @@ public class NoteController {
     @PostMapping("/patHistory/add")
     public ResponseEntity<Object> addNote(@RequestBody String body) throws JsonProcessingException {
 
-        //Convert request body to Json
+        log.info("Converting request body to Json");
         String noteJson = noteService.paramTojson(body);
-        //Convert json to Note object
+        log.info("Converting json to Note object");
         Note note = mapper.readValue(noteJson, Note.class);
         note.setCreationDate(new Date());
 
-        // validation before save
-        ResponseEntity<Object> errorResponse = noteService.getValidationErrors(note);
+        log.info("Validation before save");
+        ResponseEntity<Object> errorResponse = getValidationErrors(note);
         if (errorResponse != null) {
             return errorResponse;
         }
 
-        //TODO : Vérifie si l'id qu'on a reçu correspond à un patient
-        //endpoint menant vers
         String uri = baseUrl + portPatient + endpointPatient + "/exist/" + note.getPatId();
 
 
         Boolean patientExist = noteService.checkPatient(uri);
 
-        //si patient n'existe pas
         if (!patientExist) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -146,20 +149,18 @@ public class NoteController {
     @PutMapping("/note/{id}")
     public ResponseEntity<Object> update(@PathVariable Integer id, @RequestBody Note note) {
 
-        //Cherche person par son id
         Note noteFromDB = noteService.findNoteById(id).get();
 
         if(noteFromDB == null) {
-            //log.error("Error : Note already  exist in the Data Base.");
+            log.error("Error : Note already  exist in the Data Base.");
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        //log.debug("Set properties in the object noteFromDB.");
-        //sinon mettre à jour person
+        log.debug("Set properties in the object noteFromDB.");
         noteFromDB.setContent(note.getContent());
 
 
-        // validation before update
-        ResponseEntity<Object> errorResponse = noteService.getValidationErrors(note);
+        log.info("Validation before save");
+        ResponseEntity<Object> errorResponse = getValidationErrors(note);
         if (errorResponse != null) {
             return errorResponse;
         }
@@ -187,6 +188,24 @@ public class NoteController {
         }
     }
 
+
+    public ResponseEntity<Object> getValidationErrors(Note note) {
+        Set<ConstraintViolation<Note>> violations = validator.validate(note);
+        if (!violations.isEmpty()) {
+            AtomicInteger nb = new AtomicInteger(1);
+            JSONObject jsonError = new JSONObject();
+            violations.stream().forEach(v -> {
+                try {
+                    jsonError.put("Error " + nb.getAndIncrement(), v.getMessage());
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+            });
+            log.error(jsonError.toString());
+            return new ResponseEntity<>(jsonError.toString(), HttpStatus.BAD_REQUEST);
+        }
+        return null;
+    }
 
 }
