+++ manageNote/src/main/java/com/example/manageNote/service/NoteServiceImpl.java
@@ -3,21 +3,11 @@ package com.example.manageNote.service;
 import com.example.manageNote.model.Note;
 import com.example.manageNote.repository.NoteRepository;
 import lombok.extern.slf4j.Slf4j;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
 
 
 @Slf4j
@@ -25,14 +15,9 @@ import java.util.concurrent.atomic.AtomicInteger;
 public class NoteServiceImpl implements INoteService{
 
     private final NoteRepository noteRepository;
-    private final Validator validator;
 
-    // Injection des dépendances via le constructeur
     public NoteServiceImpl(NoteRepository noteRepository) {
         this.noteRepository = noteRepository;
-
-        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-        validator = factory.getValidator();
     }
     @Override
     public Note save(Note note) {
@@ -62,7 +47,7 @@ public class NoteServiceImpl implements INoteService{
     @Override
     public String paramTojson(String paramIn) {
         if (paramIn.startsWith("{")) {
-            //in this case it is already in json format
+            log.info("Param already in Json format");
             return paramIn;
         }
         paramIn = paramIn.replaceAll("=", "\":\"");
@@ -70,24 +55,6 @@ public class NoteServiceImpl implements INoteService{
         return "{\"" + paramIn + "\"}";
     }
 
-    @Override
-    public ResponseEntity<Object> getValidationErrors(Note note) {
-        Set<ConstraintViolation<Note>> violations = validator.validate(note);
-        if (!violations.isEmpty()) {
-            AtomicInteger nb = new AtomicInteger(1);
-            JSONObject jsonError = new JSONObject();
-            violations.stream().forEach(v -> {
-                try {
-                    jsonError.put("Error " + nb.getAndIncrement(), v.getMessage());
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-            });
-            log.error(jsonError.toString());
-            return new ResponseEntity<>(jsonError.toString(), HttpStatus.BAD_REQUEST);
-        }
-        return null;
-    }
 
     @Override
     public Boolean checkPatient(String uri) {
