--- manageNote/src/test/java/com/example/manageNote/ConfDockerServiveTest.java
@@ -1,27 +1,0 @@
-package com.example.manageNote;
-
-import com.example.manageNote.service.ConfDockerService;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import static org.junit.jupiter.api.Assertions.*;
-
-import static org.mockito.Mockito.when;
-
-public class ConfDockerServiveTest {
-    @Mock
-    ConfDockerService confDockerService;
-
-    /*@Test
-    public  void shouldReturnTrue() {
-        //Data creation
-        String activeProfile = "docker";
-
-        //Data processing
-        //On veut que la méthode isDocker retourne true/false
-        when(confDockerService.isDocker()).thenReturn(true);
-        //Test
-        //vérifie que isDocker vaut true
-        //si true on vérifie que activeProfile contient "docker"
-
-    }*/
-}
