--- manageNote/src/test/java/com/example/manageNote/ManageNoteApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.manageNote;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ManageNoteApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
